refactor: add abstraction for get cpu func
