--- cmd/main.go
@@ -1,15 +1,0 @@
-package main
-
-import (
-	"fmt"
-
-	"github.com/codescalersinternships/psutil-go-DohaElsawy/psutil"
-)
-
-
-func main() {
-	c := psutil.InitCPU()
-	c.GetCpuInfo()
-
-	fmt.Printf(c.String())
-}
